refactor(tests): Extract inline tests to separate files
- Create playlist_runner_tests.rs from inline tests in playlist_runner.rs
- Create keyboard_tests.rs from inline tests in keyboard.rs
- Create runner_tests.rs from inline tests in runner/mod.rs
- Update src/tests/mod.rs to include new test modules
- Update DEVELOPMENT_GUIDE.md to clarify test organization pattern:
  - All tests MUST be in separate files, not inline
  - Added inline tests to "Common Mistakes" section
