+++ src/tests/mod.rs
@@ -6,10 +6,13 @@ mod cli_output_integration_tests;
 mod cli_setup_tests;
 mod config_tests;
 mod error_tests;
+mod keyboard_tests;
 mod lock_tests;
 mod logging_tests;
+mod playlist_runner_tests;
 mod playlist_tests;
 mod process_control_tests;
+mod runner_tests;
 mod runtime_state_tests;
 mod schedule_runner_tests;
 mod scheduler_tests;
